+++ FD502/fd502.cpp
@@ -405,7 +405,7 @@ void Load_Disk(unsigned char disk)
         {
           FileNotSelected = 0;
           string tmp = ofn.lpstrFile;
-          int idx;
+          size_t idx;
           idx = tmp.find_last_of("\\");
           tmp = tmp.substr(0, idx);
           strcpy(FloppyPath, tmp.c_str());
